Refactor 'MAuthCoreExtensions.cs'  and implement using interface
