+++ src/Medidata.MAuth.Core/IMAuthCoreImplementation.cs
@@ -1,0 +1,14 @@
+﻿using System.Net.Http;
+using System.Threading.Tasks;
+
+namespace Medidata.MAuth.Core
+{
+    internal interface IMAuthCoreImplementation
+    {
+        Task<HttpRequestMessage> Sign(HttpRequestMessage request, MAuthSigningOptions options);
+
+        bool Verify(byte[] signedData, byte[] signature, string publicKey);
+
+        Task<byte[]> GetSignature(HttpRequestMessage request, AuthenticationInfo authInfo);
+    }
+}
