+++ tests/Medidata.MAuth.Tests/MAuthAuthenticatorTests.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Linq;
 using System.Net;
+using System.Net.Http;
 using System.Threading.Tasks;
 using Medidata.MAuth.Core;
 using Medidata.MAuth.Tests.Infrastructure;
@@ -40,9 +41,10 @@ namespace Medidata.MAuth.Tests
             var testData = await method.FromResource();
 
             var authenticator = new MAuthAuthenticator(TestExtensions.ServerOptions);
+            var mAuthCore = MAuthCoreImplementation.MAuthCore;
 
-            var signedRequest = await testData.ToHttpRequestMessage()
-                .AddAuthenticationInfo(new PrivateKeyAuthenticationInfo()
+            var signedRequest = await mAuthCore
+                .AddAuthenticationInfo(testData.ToHttpRequestMessage(), new PrivateKeyAuthenticationInfo()
                 {
                     ApplicationUuid = testData.ApplicationUuid,
                     PrivateKey = TestExtensions.ClientPrivateKey,
@@ -70,9 +72,10 @@ namespace Medidata.MAuth.Tests
 
             var authenticator = new MAuthAuthenticator(TestExtensions.GetServerOptionsWithAttempts(
                 policy, shouldSucceedWithin: true));
+            var mAuthCore = MAuthCoreImplementation.MAuthCore;
 
-            var signedRequest = await testData.ToHttpRequestMessage()
-                .AddAuthenticationInfo(new PrivateKeyAuthenticationInfo()
+            var signedRequest = await mAuthCore
+                .AddAuthenticationInfo(testData.ToHttpRequestMessage(),new PrivateKeyAuthenticationInfo()
                 {
                     ApplicationUuid = testData.ApplicationUuid,
                     PrivateKey = TestExtensions.ClientPrivateKey,
@@ -99,9 +102,10 @@ namespace Medidata.MAuth.Tests
 
             var authenticator = new MAuthAuthenticator(TestExtensions.GetServerOptionsWithAttempts(
                 policy, shouldSucceedWithin: false));
+            var mAuthCore = MAuthCoreImplementation.MAuthCore;
 
-            var signedRequest = await testData.ToHttpRequestMessage()
-                .AddAuthenticationInfo(new PrivateKeyAuthenticationInfo()
+            var signedRequest = await mAuthCore
+                .AddAuthenticationInfo(testData.ToHttpRequestMessage(),new PrivateKeyAuthenticationInfo()
                 {
                     ApplicationUuid = testData.ApplicationUuid,
                     PrivateKey = TestExtensions.ClientPrivateKey,
@@ -130,9 +134,10 @@ namespace Medidata.MAuth.Tests
             // Arrange
             var testData = await method.FromResource();
             var expectedMAuthHeader = testData.MAuthHeader;
+            var mAuthCore = MAuthCoreImplementation.MAuthCore;
 
             // Act
-            var actual = await testData.ToHttpRequestMessage().Sign(TestExtensions.ClientOptions(testData.SignedTime));
+            var actual = await mAuthCore.Sign(testData.ToHttpRequestMessage(),TestExtensions.ClientOptions(testData.SignedTime));
 
             // Assert
             Assert.Equal(expectedMAuthHeader, actual.Headers.GetFirstValueOrDefault<string>(Constants.MAuthHeaderKey));
@@ -140,6 +145,31 @@ namespace Medidata.MAuth.Tests
                 testData.SignedTime.ToUnixTimeSeconds(),
                 actual.Headers.GetFirstValueOrDefault<long>(Constants.MAuthTimeHeaderKey)
             );
+        }
+
+        [Theory]
+        [InlineData("GET")]
+        [InlineData("DELETE")]
+        [InlineData("POST")]
+        [InlineData("PUT")]
+        public static async Task GetAuthenticationInfo_WithSignedRequest_WillReturnCorrectAuthInfo(string method)
+        {
+            // Arrange
+            var authenticator = new MAuthAuthenticator(TestExtensions.ServerOptions);
+            var testData = await method.FromResource();
+            var request = new HttpRequestMessage(new HttpMethod(testData.Method), TestExtensions.TestUri);
+
+            request.Headers.Add(
+                Constants.MAuthHeaderKey, testData.MAuthHeader);
+            request.Headers.Add(Constants.MAuthTimeHeaderKey, testData.SignedTimeUnixSeconds.ToString());
+
+            // Act
+            var actual = authenticator.GetAuthenticationInfo(request);
+
+            // Assert
+            Assert.Equal(testData.ApplicationUuid, actual.ApplicationUuid);
+            Assert.Equal(Convert.FromBase64String(testData.Payload), actual.Payload);
+            Assert.Equal(testData.SignedTime, actual.SignedTime);
         }
     }
 }
