+++ version.props
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project>
   <PropertyGroup>
-    <Version>3.1.2</Version>
+    <Version>3.1.3</Version>
   </PropertyGroup>
 </Project>
